refactor(core): move the account-type taxonomy into rustledger-core (#1402)
`ACCOUNT_TYPES` + `account_type` (the five Beancount root types and the
lowercased-root classifier) were pinned in the `rustledger-ffi-wasi` FFI crate.
They are a core domain concept, so relocate them to `rustledger-core`
(`identifiers`, next to `Account`/`is_subaccount_or_equal`) as the single
source of truth, and re-export from `ffi-wasi::helpers` so the existing FFI
call sites (`util.types`, `util.getAccountType`) are unchanged.

Behavior is identical (these remain the default English roots; the `name_*`
loader options that can rename them per-ledger are still not modeled — noted in
the doc). Added a unit test in core.

Refs #1402.
