+++ crates/rustledger-core/src/lib.rs
@@ -78,7 +78,9 @@ pub use format::{
     Alignment, FormatConfig, FormatLine, format_directive_lines, format_directives,
     format_posting_line, posting_format_line, render_lines, resolve_alignment,
 };
-pub use identifiers::{Account, Currency, Link, Tag, is_subaccount_or_equal};
+pub use identifiers::{
+    ACCOUNT_TYPES, Account, Currency, Link, Tag, account_type, is_subaccount_or_equal,
+};
 pub use implicit_prices::extract_per_unit_price;
 pub use intern::{InternedStr, StringInterner};
 pub use inventory::{
